Parameterize asserts and use decorator for test client login
